feat: extract superuser console into dedicated /superuser route (#805)
Moves all superuser-only sub-tabs (EULA, Storage, CVE, Workers, Deletion,
Reconcile, Maintenance, Scheduled Tasks) out of AdminPage into a new
SuperuserPage at /superuser/:section with URL-based navigation. Adds
Superuser Console as a top-level sidebar nav item visible only to superusers,
and updates ProtectedRoute with requireSuperuser guard.
