+++ frontend/src/lib/icons.ts
@@ -35,6 +35,7 @@ import {
   Settings,
   Share2,
   ShieldCheck,
+  Terminal,
   Wrench,
   X,
   Zap,
@@ -61,6 +62,7 @@ export const AppIcons = {
   collections: BookOpen,
   settings: Settings,
   admin: ShieldCheck,
+  superuser: Terminal,
   logout: LogOut,
   feedback: MessageSquare,
   onboarding: ListChecks,
